Split up JS into modules.
