+++ js_src/utils/Loading.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class Loading extends React.Component {
+    render() {
+        return (
+            <div className="loading">Loading</div>
+        );
+    }
+}
+
+export default Loading;
