+++ js_src/utils/NoMatch.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class NoMatch extends React.Component {
+    render() {
+        return (
+            <h2>Oops - Page not found</h2>
+        );
+    }
+}
+
+export default NoMatch;
